- Copied V3.3 upgrade scripts into V3.5

+++ sql/postgresql/upgrade/upgrade-3.1.3.0.0-3.2.0.0.0.sql
@@ -1,16 +1,50 @@
+-- upgrade-3.1.3.0.0-3.2.0.0.0.sql
+
+SELECT acs_log__debug('/packages/intranet-timesheet2-invoices/sql/postgresql/upgrade/upgrade-3.1.3.0.0-3.2.0.0.0.sql','');
 
 
--- Rename the im_price_idx to im_timesheet_price_idx
-drop index im_price_idx;
+
+create or replace function inline_1 ()
+returns integer as '
+declare
+        v_count                 integer;
+begin
+	select count(*) into v_count from pg_indexes
+	where indexname = ''im_price_idx'';
+	if v_count = 0 then return 0; end if;
+
+	-- Rename the im_price_idx to im_timesheet_price_idx
+	drop index im_price_idx;
+
+	return 0;
+end;' language 'plpgsql';
+select inline_1 ();
+drop function inline_1();
 
 
-create unique index im_timesheet_price_idx on im_timesheet_prices (
-        uom_id, company_id, task_type_id, material_id, currency
-);
+
+create or replace function inline_1 ()
+returns integer as '
+declare
+        v_count                 integer;
+begin
+	select count(*) into v_count from pg_indexes
+	where indexname = ''im_timesheet_price_idx'';
+	if v_count > 0 then return 0; end if;
+
+	create unique index im_timesheet_price_idx on im_timesheet_prices (
+	        uom_id, company_id, task_type_id, material_id, currency
+	);
+
+	return 0;
+end;' language 'plpgsql';
+select inline_1 ();
+drop function inline_1();
+
+
 
 
 -- Recompile the __delete procedure
-
 -- Delete a single invoice (if we know its ID...)
 create or replace function  im_timesheet_invoice__delete (integer)
 returns integer as '
